rfr: rename cta-notification to cta-snackbar
TRACEFOSS-685
